@@ -11,15 +11,16 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "bristol-su/support": "^4.0",
+        "php": "^8.0",
+        "bristol-su/support": "dev-portal-v4 as 4.0",
         "watr/typeform": "^1.5",
-        "bristol-su/typeform-service": "^1.0.1",
+        "bristol-su/typeform-service": "dev-feature/port-modules as 1.0.1",
         "ext-json": "*",
         "spatie/laravel-rate-limited-job-middleware": "^1.4"
     },
     "require-dev": {
-        "orchestra/testbench": "^4.0",
+        "orchestra/testbench": "^6.19",
+        "laravel/legacy-factories": "^1.1",
         "phpspec/prophecy-phpunit": "^2.0"
     },
     "autoload": {
@@ -32,6 +33,7 @@
             "BristolSU\\Module\\Tests\\Typeform\\": "tests/"
         }
     },
+    "repositories": [{"type":  "path", "url":  "./../typeform-service"}],
     "extra": {
         "laravel": {
             "providers": [
