@@ -16,7 +16,7 @@ class Answer extends Model
     protected $casts = [
         'encoded' => 'boolean'
     ];
-    
+
     public function setAnswerAttribute($answer)
     {
         if(is_array($answer)) {
@@ -44,5 +44,16 @@ class Answer extends Model
     {
         return $this->belongsTo(Response::class);
     }
-    
+
+    /**
+     * Prepare a date for array / JSON serialization.
+     *
+     * @param  \DateTimeInterface  $date
+     * @return string
+     */
+    protected function serializeDate(\DateTimeInterface $date)
+    {
+        return $date->format('Y-m-d H:i:s');
+    }
+
 }
