@@ -6,11 +6,11 @@ use Illuminate\Database\Eloquent\Model;
 
 class Field extends Model
 {
-    
+
     protected $keyType = 'string';
-    
+
     public $incrementing = false;
-    
+
     protected $table = 'typeform_fields';
 
     protected $fillable = [
@@ -20,6 +20,17 @@ class Field extends Model
     public function answers()
     {
         return $this->hasMany(Answer::class);
+    }
+
+    /**
+     * Prepare a date for array / JSON serialization.
+     *
+     * @param  \DateTimeInterface  $date
+     * @return string
+     */
+    protected function serializeDate(\DateTimeInterface $date)
+    {
+        return $date->format('Y-m-d H:i:s');
     }
 
 }
